+++ backend-v3/libs/aqKumabank/index.js
@@ -1,0 +1,6 @@
+'use strict'
+
+const constants = require('./constants')
+
+module.exports.constants = constants
+module.exports.responseCodes = constants.responseCodes