+++ backend-v3/libs/trxManager/handlers/kumabank.js
@@ -8,6 +8,16 @@ const _ = require('lodash')
 const { createError } = require('libs/error')
 const timer = require('libs/timer')
 const card = require('libs/card')
+const kumabank = require('libs/aqKumabank')
+
+function getResponseMessage (responseCode) {
+  if (!(responseCode in kumabank.responseCodes)) responseCode = '00'
+  const handlerResponse = {
+    responseCode,
+    responseMessage: kumabank.responseCodes[responseCode]
+  }
+  return handlerResponse
+}
 
 const {
   errorTypes,
@@ -32,7 +42,9 @@ module.exports = {
   properties: {
     flows: [
       transactionFlows.GET_TOKEN,
-      transactionFlows.VOID
+      transactionFlows.VOID,
+      transactionFlows.REVERSE_SUCCESS,
+      transactionFlows.REVERSE_VOID
     ],
     tokenTypes: [],
     userTokenTypes: [
@@ -41,6 +53,8 @@ module.exports = {
   },
   async handler (ctx) {
     // const debug = require('debug')('mika:trxManager:kumabank:handler')
+
+    ctx.transaction.status = transactionStatuses.PROCESSING
 
     const userToken = ctx.userToken
 
@@ -70,6 +84,15 @@ module.exports = {
       })
     }
 
+    if (
+      userToken.pin ||
+      userToken.pinBlock ||
+      userToken.standardPinBlock ||
+      userToken.doubleLengthPinBlock
+    ) {
+      userToken.signature = undefined
+    }
+
     let emvTags
     let track2
     if (userToken.emv) {
@@ -79,17 +102,24 @@ module.exports = {
       track2 = userToken.track2
     }
 
+    let entryMode
+    if (userToken.emv) {
+      if (userToken.signature) {
+        entryMode = card.posEntryMode.CHIP_WITHOUT_PIN
+      } else {
+        entryMode = card.posEntryMode.CHIP_WITH_PIN
+      }
+    } else if (userToken.track2) {
+      if (userToken.signature) {
+        entryMode = card.posEntryMode.MAGSTRIPE_WITHOUT_PIN
+      } else {
+        entryMode = card.posEntryMode.MAGSTRIPE_WITH_PIN
+      }
+    }
+
     const track2Component = card.track2GetComponent(track2)
-
-    ctx.transaction.status = transactionStatuses.SUCCESS
-    ctx.transaction.reference = String(Math.floor(1 + Math.random() * 999999999998)).padStart(12, '0')
-    ctx.transaction.referenceName = 'rrn'
-    ctx.transaction.customerReference = track2Component.truncatedPan
-    ctx.transaction.customerReferenceName = 'truncatedPan'
-    ctx.transaction.authorizationReference = String(Math.floor(1 + Math.random() * 999998)).padStart(6, '0')
-    ctx.transaction.authorizationReferenceName = 'approvalCode'
-
     ctx.transaction.properties.iin = track2Component.first6
+
     if (ctx.local.cardIin) {
       ctx.transaction.properties.cardIin = {
         name: ctx.local.cardIin.name,
@@ -98,35 +128,167 @@ module.exports = {
         type: ctx.local.cardIin.cardTypeId
       }
     }
+    ctx.transaction.properties.posCode = '00'
+    ctx.transaction.properties.entryMode = entryMode
+    ctx.transaction.properties.signature = userToken.signature
 
-    /*
-    ctx.transaction.encryptedProperties.track2Component = {
-      pan: track2Component.pan,
-      expirationDate: track2Component.expirationDate,
-      serviceCode: track2Component.serviceCode
+    ctx.debug = {
+      createDelay: 0,
+      createNoResponse: false,
+      createResponseCode: '00',
+      ...ctx.debug
     }
-    ctx.transaction.changed('encryptedProperties', 'true')
-    */
 
-    // Random delay
-    await timer.delay(100 + Math.random() * 5000)
-
-    // Random failure
-    /*
-    if (Math.random() < 0.01) {
-      throw createError({
-        name: errorTypes.ACQUIRER_HOST_NO_RESPONSE
-      })
+    if (ctx.debug.createDelay) {
+      await timer.delay(ctx.debug.createDelay)
+    } else {
+      await timer.delay(100 + Math.random() * 2000)
     }
-    */
+
+    if (ctx.debug.createNoResponse) {
+      ctx.handlerResponse = undefined
+      return
+    }
+
+    if (ctx.debug.createPinError) {
+      ctx.handlerResponse = getResponseMessage('55')
+    } else {
+      ctx.handlerResponse = getResponseMessage(ctx.debug.createResponseCode)
+    }
+
+    if (ctx.handlerResponse.responseCode === '00') {
+      ctx.transaction.status = transactionStatuses.SUCCESS
+
+      ctx.transaction.reference = String(Math.floor(1 + Math.random() * 999999999998)).padStart(12, '0')
+      ctx.transaction.referenceName = 'rrn'
+
+      ctx.transaction.customerReference = track2Component.truncatedPan
+      ctx.transaction.customerReferenceName = 'truncatedPan'
+
+      ctx.transaction.authorizationReference = String(Math.floor(1 + Math.random() * 999998)).padStart(6, '0')
+      ctx.transaction.authorizationReferenceName = 'approvalCode'
+
+      ctx.transaction.acquirerTimeAt = (new Date()).toISOString()
+    } else {
+      ctx.handlerFailed = true
+      ctx.transaction.status = transactionStatuses.FAILED
+    }
   },
   async voidHandler (ctx) {
     // const debug = require('debug')('mika:trxManager:kumabank:voidHandler')
-    ctx.transaction.status = transactionStatuses.VOIDED
+
+    ctx.transaction.status = transactionStatuses.VOIDING
+
+    ctx.debug = {
+      voidDelay: 0,
+      voidNoResponse: false,
+      voidResponseCode: '00',
+      ...ctx.debug
+    }
+
+    if (ctx.debug.voidDelay) {
+      await timer.delay(ctx.debug.voidDelay)
+    } else {
+      await timer.delay(100 + Math.random() * 2000)
+    }
+
+    if (ctx.debug.voidNoResponse) {
+      ctx.handlerResponse = undefined
+      return
+    }
+
+    ctx.handlerResponse = getResponseMessage(ctx.debug.voidResponseCode)
+
+    if (ctx.handlerResponse.responseCode === '00') {
+      ctx.transaction.status = transactionStatuses.SUCCESS
+
+      ctx.transaction.voidReference = String(Math.floor(1 + Math.random() * 999999999998)).padStart(12, '0')
+      ctx.transaction.voidReferenceName = 'rrn'
+
+      ctx.transaction.voidAuthorizationReference = String(Math.floor(1 + Math.random() * 999998)).padStart(6, '0')
+      ctx.transaction.voidAuthorizationReferenceName = 'approvalCode'
+
+      ctx.transaction.voidAcquirerTimeAt = (new Date()).toISOString()
+    }
+  },
+  async reverseHandler (ctx) {
+    // const debug = require('debug')('mika:trxManager:kumabank:reverseHandler')
+
+    ctx.debug = {
+      reverseDelay: 0,
+      reverseNoResponse: false,
+      reverseResponseCode: '00',
+      ...ctx.debug
+    }
+
+    if (ctx.debug.reverseDelay) {
+      await timer.delay(ctx.debug.reverseDelay)
+    } else {
+      await timer.delay(100 + Math.random() * 2000)
+    }
+
+    if (ctx.debug.reverseNoResponse) {
+      ctx.handlerResponse = undefined
+      return
+    }
+
+    ctx.handlerResponse = getResponseMessage(ctx.debug.reverseResponseCode)
+
+    if (ctx.handlerResponse.responseCode === '00') {
+      ctx.transaction.status = transactionStatuses.REVERSED
+    }
+  },
+  async reverseVoidHandler (ctx) {
+    ctx.debug = {
+      reverseVoidDelay: 0,
+      reverseVoidNoResponse: false,
+      reverseVoidResponseCode: '00',
+      ...ctx.debug
+    }
+
+    if (ctx.debug.reverseVoidDelay) {
+      await timer.delay(ctx.debug.reverseVoidDelay)
+    } else {
+      await timer.delay(100 + Math.random() * 2000)
+    }
+
+    if (ctx.debug.reverseVoidNoResponse) {
+      ctx.handlerResponse = undefined
+      return
+    }
+
+    ctx.handlerResponse = getResponseMessage(ctx.debug.reverseVoidResponseCode)
+
+    if (ctx.handlerResponse.responseCode === '00') {
+      ctx.transaction.status = transactionStatuses.SUCCESS
+    }
   },
   async agentSettleHandler (ctx) {
     // const debug = require('debug')('mika:trxManager:kumabank:agentSettleHandler')
-    ctx.settleBatch.status = settleBatchStatuses.CLOSED
-    ctx.settleBatch.acquirerTimeAt = (new Date()).toISOString()
+
+    ctx.debug = {
+      agentSettleDelay: 0,
+      agentSettleNoResponse: false,
+      agentSettleResponseCode: '00',
+      ...ctx.debug
+    }
+
+    if (ctx.debug.agentSettleDelay) {
+      await timer.delay(ctx.debug.agentSettleDelay)
+    } else {
+      await timer.delay(100 + Math.random() * 2000)
+    }
+
+    if (ctx.debug.agentSettleNoResponse) {
+      ctx.handlerResponse = undefined
+      return
+    }
+
+    ctx.handlerResponse = getResponseMessage(ctx.debug.agentSettleResponseCode)
+
+    if (ctx.handlerResponse.responseCode === '00') {
+      ctx.settleBatch.status = settleBatchStatuses.CLOSED
+      ctx.settleBatch.acquirerTimeAt = (new Date()).toISOString()
+    }
   }
 }